Navigation수정 & Fonts 추가
