+++ src/containers/group/detail/index.js
@@ -6,7 +6,7 @@ import Search from './components/search';
 import Bottom from './components/bottom';
 import ZeroTiccle from './components/zeroTiccle';
 
-export default function GroupDetail(){
+const GroupDetail = () => {
   return(
     <>
       <GroupInfo title={"현판"} imgUrl={'https://t1.daumcdn.net/thumb/R720x0.fpng/?fname=http://t1.daumcdn.net/brunch/service/user/8fXh/image/CyKAu5r6yUDSnRAy28UDlDEpCDs.png'} content={"데못죽 같은 거 모아두는"}/>
@@ -37,3 +37,5 @@ const styles = StyleSheet.create({
     height: 60,
   },
 })
+
+export default GroupDetail;
